requirements.txt
pandas>=1.1.5
requests>=2.25.1
apache-airflow-providers-google>=8.1.0
